+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-Unreleased
-----------
+0.1.7
+-----
 - Added `RefValMut` providing advanced mapping functionality for mutable
   borrows
 - Removed implicit `T: Clone` constraint from `RefVal::map`